Switch to new column indices-based get_schema
